test(ui): Move provider and hook logic tests out of source files
