+++ ui/src/providers/home-data/favorites.ts
@@ -120,91 +120,3 @@ export const toggleFavoriteItem = (
 
   return addFavoriteItem(favorites, favorite, now);
 };
-
-if (import.meta.vitest) {
-  const { describe, expect, it } = import.meta.vitest;
-
-  describe('favorite helpers', () => {
-    const now = new Date('2026-06-16T12:00:00.000Z');
-
-    const favorite: FavoriteItemInput = {
-      id: buildOrganizationFavoriteId(1),
-      type: 'organization',
-      name: 'Acme',
-      breadcrumbs: ['Acme'],
-      to: '/organizations/$orgId',
-      params: { orgId: '1' },
-    };
-
-    it('normalizes favorite group order', () => {
-      expect(
-        normalizeFavoriteGroupOrder(['product', 'run', 'product'])
-      ).toStrictEqual(['product', 'run', 'repository', 'organization']);
-    });
-
-    it('builds stable favorite IDs', () => {
-      expect(buildOrganizationFavoriteId(123)).toBe('organization:123');
-      expect(buildProductFavoriteId(123, 456)).toBe('product:123:456');
-      expect(buildRepositoryFavoriteId(123, 456, 789)).toBe(
-        'repository:123:456:789'
-      );
-      expect(buildRunFavoriteId(42)).toBe('run:42');
-    });
-
-    it('adds favorites newest first', () => {
-      expect(addFavoriteItem([], favorite, now)).toStrictEqual([
-        { ...favorite, starredAt: now.toISOString() },
-      ]);
-    });
-
-    it('de-duplicates favorites by ID', () => {
-      const first = addFavoriteItem([], favorite, now);
-      const updated = addFavoriteItem(
-        first,
-        { ...favorite, name: 'Acme Updated' },
-        new Date('2026-06-16T13:00:00.000Z')
-      );
-
-      expect(updated).toStrictEqual([
-        {
-          ...favorite,
-          name: 'Acme Updated',
-          starredAt: '2026-06-16T13:00:00.000Z',
-        },
-      ]);
-    });
-
-    it('updates a favorite in place', () => {
-      const favorites = addFavoriteItem([], favorite, now);
-
-      expect(
-        updateFavoriteItem(favorites, {
-          ...favorite,
-          name: 'Acme Updated',
-          starredAt: now.toISOString(),
-        })
-      ).toStrictEqual([
-        {
-          ...favorite,
-          name: 'Acme Updated',
-          starredAt: now.toISOString(),
-        },
-      ]);
-    });
-
-    it('removes favorites by ID', () => {
-      const favorites = addFavoriteItem([], favorite, now);
-
-      expect(removeFavoriteItem(favorites, favorite.id)).toStrictEqual([]);
-    });
-
-    it('toggles favorites', () => {
-      const favorites = toggleFavoriteItem([], favorite, now);
-
-      expect(favorites).toStrictEqual([
-        { ...favorite, starredAt: now.toISOString() },
-      ]);
-      expect(toggleFavoriteItem(favorites, favorite, now)).toStrictEqual([]);
-    });
-  });
-}
